initial move of core code
