+++ ethercat_driver/include/ethercat_driver/ethercat_driver.hpp
@@ -19,7 +19,6 @@
 #include <memory>
 #include <string>
 #include <vector>
-#include <pluginlib/class_loader.hpp>
 #include "hardware_interface/handle.hpp"
 #include "hardware_interface/hardware_info.hpp"
 #include "hardware_interface/system_interface.hpp"
@@ -27,10 +26,8 @@
 #include "rclcpp/macros.hpp"
 #include "rclcpp_lifecycle/node_interfaces/lifecycle_node_interface.hpp"
 #include "rclcpp_lifecycle/state.hpp"
+#include "ethercat_driver/ethercat_bus_manager.hpp"
 #include "ethercat_driver/visibility_control.h"
-#include "ethercat_interface/ec_slave.hpp"
-#include "ethercat_interface/ec_master.hpp"
-#include "yaml-cpp/yaml.h"
 
 using CallbackReturn = rclcpp_lifecycle::node_interfaces::LifecycleNodeInterface::CallbackReturn;
 
@@ -68,49 +65,6 @@ public:
   hardware_interface::return_type write(const rclcpp::Time &, const rclcpp::Duration &) override;
 
 protected:
-  std::vector<std::unordered_map<std::string, std::string>> getEcModuleParam(
-    const std::string & urdf,
-    const std::string & component_name,
-    const std::string & component_type);
-
-  uint16_t getAliasOrDefaultAlias(
-    const std::unordered_map<std::string,
-    std::string> & slave_parameters);
-
-  virtual CallbackReturn setupMaster();
-
-  CallbackReturn configNetwork();
-
-  /** @brief Load transfer config YAML file
-   * One use case is to load transfers for FailSafe Over EtherCAT Safety
-   * @param[out] node YAML node containing the transfer configuration root
-   * @param[in] path Path to the YAML file, if empty, the file is loaded from the *fsoe_config*
-   * or *transfer_config* of the YAML document
-   */
-  void loadTransferConfigYamlFile(YAML::Node & node, const std::string & path = "");
-
-  /** @brief Get transfer module parameters from YAML file
-   * @param[in] config YAML node containing the transfer configuration root
-   * @return Vector of maps containing transfer module parameters, each map corresponds to a module
-   * involved in a transfer
-   */
-  std::vector<std::unordered_map<std::string, std::string>> getEcTransferModuleParam(
-    const YAML::Node & config);
-
-  /** @brief Get transfer nets from YAML file
-   * @param[in] config YAML node containing the transfer configuration root
-   * @return Vector of transfer nets
-   */
-  std::vector<ethercat_interface::EcTransferNet> getEcTransferNets(const YAML::Node & config);
-
-  /** @brief Configure the transfer networks
-   */
-  void configTransferNetwork();
-
-protected:
-  std::vector<std::shared_ptr<ethercat_interface::EcSlave>> ec_modules_;
-  std::vector<std::unordered_map<std::string, std::string>> ec_module_parameters_;
-
   std::vector<std::vector<double>> hw_joint_commands_;
   std::vector<std::vector<double>> hw_sensor_commands_;
   std::vector<std::vector<double>> hw_gpio_commands_;
@@ -118,25 +72,7 @@ protected:
   std::vector<std::vector<double>> hw_sensor_states_;
   std::vector<std::vector<double>> hw_gpio_states_;
 
-  pluginlib::ClassLoader<ethercat_interface::EcSlave> ec_loader_{
-    "ethercat_interface", "ethercat_interface::EcSlave"};
-
-  double control_frequency_;
-
-  std::shared_ptr<ethercat_interface::EcMaster> master_;
-  std::mutex ec_mutex_;
-  bool activated_;
-
-  /** Transfer nets */
-  std::vector<ethercat_interface::EcTransferNet> ec_transfer_nets_;
-
-  /** Indexes of modules inside ec_modules_ vector that are transfer masters */
-  std::vector<size_t> ec_transfer_masters_;
-  /** Indexes of modules inside ec_modules_ vector that are transfer slaves only */
-  std::vector<size_t> ec_transfer_slaves_;
-
-  /** Empty interfaces */
-  std::vector<double> empty_interface_;
+  EthercatBusManager bus_manager_;
 };
 }  // namespace ethercat_driver
