+++ ethercat_driver/src/ethercat_ros2_control_xml_parser.cpp
@@ -1,0 +1,85 @@
+// Copyright 2022 ICUBE Laboratory, University of Strasbourg
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "ethercat_driver/ethercat_ros2_control_xml_parser.hpp"
+
+#include <stdexcept>
+#include <string>
+
+#include <tinyxml2.h>
+
+namespace ethercat_driver
+{
+
+std::vector<std::unordered_map<std::string, std::string>>
+EthercatRos2ControlXmlParser::getEcModuleParam(
+  const std::string & urdf,
+  const std::string & component_name,
+  const std::string & component_type)
+{
+  // Check if everything OK with URDF string
+  if (urdf.empty()) {
+    throw std::runtime_error("empty URDF passed to robot");
+  }
+  tinyxml2::XMLDocument doc;
+  if (!doc.Parse(urdf.c_str()) && doc.Error()) {
+    throw std::runtime_error("invalid URDF passed in to robot parser");
+  }
+  if (doc.Error()) {
+    throw std::runtime_error("invalid URDF passed in to robot parser");
+  }
+
+  tinyxml2::XMLElement * robot_it = doc.RootElement();
+  if (std::string("robot").compare(robot_it->Name())) {
+    throw std::runtime_error("the robot tag is not root element in URDF");
+  }
+
+  const tinyxml2::XMLElement * ros2_control_it = robot_it->FirstChildElement("ros2_control");
+  if (!ros2_control_it) {
+    throw std::runtime_error("no ros2_control tag");
+  }
+
+  std::vector<std::unordered_map<std::string, std::string>> module_params;
+  std::unordered_map<std::string, std::string> module_param;
+
+  while (ros2_control_it) {
+    const auto * ros2_control_child_it = ros2_control_it->FirstChildElement(component_type.c_str());
+    while (ros2_control_child_it) {
+      if (!component_name.compare(ros2_control_child_it->Attribute("name"))) {
+        const auto * ec_module_it = ros2_control_child_it->FirstChildElement("ec_module");
+        while (ec_module_it) {
+          module_param.clear();
+          module_param["name"] = ec_module_it->Attribute("name");
+          const auto * plugin_it = ec_module_it->FirstChildElement("plugin");
+          if (NULL != plugin_it) {
+            module_param["plugin"] = plugin_it->GetText();
+          }
+          const auto * param_it = ec_module_it->FirstChildElement("param");
+          while (param_it) {
+            module_param[param_it->Attribute("name")] = param_it->GetText();
+            param_it = param_it->NextSiblingElement("param");
+          }
+          module_params.push_back(module_param);
+          ec_module_it = ec_module_it->NextSiblingElement("ec_module");
+        }
+      }
+      ros2_control_child_it = ros2_control_child_it->NextSiblingElement(component_type.c_str());
+    }
+    ros2_control_it = ros2_control_it->NextSiblingElement("ros2_control");
+  }
+
+  return module_params;
+}
+
+}  // namespace ethercat_driver
